+++ pb/pb_migrations/1746955376_updated_tickets.js
@@ -1,0 +1,17 @@
+/// <reference path="../pb_data/types.d.ts" />
+migrate((app) => {
+  const collection = app.findCollectionByNameOrId("pbc_3306545694");
+
+  // remove field
+  collection.fields.removeById("text2929936659");
+
+  // update collection data
+  unmarshal(
+    {
+      updateRule: '@request.auth.collectionName = "agents"',
+    },
+    collection,
+  );
+
+  return app.save(collection);
+});