+++ src/features/tickets/TicketList.tsx
@@ -1,12 +1,9 @@
-import { Card, CardContent } from "@/components/ui/card";
-import { useLocation } from "wouter";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { cn } from "@/lib/utils";
-import { pb } from "@/lib/db/pocketbase";
 import { useCheckRole } from "../auth/useCheckRole";
 import { TicketStatus } from "./ticket.types";
 import { useFilterTickets } from "./useFilterTickets";
 import CreateTicketButtonDialog from "./CreateTicketButtonDialog";
+import TicketListItem from "./TicketListItem";
 
 type Props = {
   status?: TicketStatus;
@@ -15,7 +12,6 @@ type Props = {
 
 export default function TicketList(props: Props) {
   const { className, status } = props;
-  const [, navigate] = useLocation();
   const isCustomer = useCheckRole("customers");
   const filteredTickets = useFilterTickets();
   let tickets = filteredTickets.all;
@@ -41,38 +37,7 @@ export default function TicketList(props: Props) {
   return (
     <div className={cn("flex flex-col gap-4", className)}>
       {tickets.map((ticket) => (
-        <Card
-          key={ticket.id}
-          className="cursor-pointer px-2 py-1"
-          onClick={() => navigate(`/ticket/${ticket.id}`)}
-        >
-          <CardContent className="flex flex-col gap-2 p-2">
-            <div className="border-muted flex items-center justify-between gap-2 border-b pb-2 text-xs">
-              <div className="flex items-center gap-2">
-                <Avatar>
-                  <AvatarImage
-                    src={`https://api.dicebear.com/9.x/identicon/svg?seed=${ticket.expand?.customer_id?.fullName.slice(0, 4)}`}
-                  />
-                  <AvatarFallback>
-                    {ticket.expand?.customer_id?.fullName.slice(0, 2)}
-                  </AvatarFallback>
-                </Avatar>
-
-                <span className="font-bold">
-                  {ticket.customer_id === pb.authStore?.record?.id
-                    ? "You"
-                    : ticket.expand?.customer_id?.fullName || ""}
-                </span>
-              </div>
-
-              <span className="text-muted-foreground">
-                {new Date(ticket.created).toLocaleString()}
-              </span>
-            </div>
-
-            <div className="whitespace-pre-wrap">{ticket.message}</div>
-          </CardContent>
-        </Card>
+        <TicketListItem key={ticket.id} ticket={ticket} />
       ))}
     </div>
   );