+++ src/features/tickets/useTicketStore.ts
@@ -6,6 +6,7 @@ type TicketState = {
   setTickets: (tickets: TicketExpand[]) => void;
   addTicket: (ticket: TicketExpand) => void;
   removeTicket: (id: string) => void;
+  updateTicket: (ticket: TicketExpand) => void;
 };
 
 const initState: Pick<TicketState, "tickets"> = {
@@ -19,6 +20,10 @@ const useTicketStore = create<TicketState>((set) => ({
     set((state) => ({ tickets: [ticket, ...state.tickets] })),
   removeTicket: (id) =>
     set((state) => ({ tickets: state.tickets.filter((t) => t.id !== id) })),
+  updateTicket: (ticket) =>
+    set((state) => ({
+      tickets: state.tickets.map((t) => (t.id === ticket.id ? ticket : t)),
+    })),
 }));
 
 export default useTicketStore;